% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeapYear.R
\name{LeapYear}
\alias{LeapYear}
\title{Checks Whether A Year Is Leap Year}
\usage{
LeapYear(year)
}
\arguments{
\item{year}{A numeric value indicating the year in the Gregorian calendar.}
}
\value{
Boolean telling whether the year is a leap year or not.
}
\description{
This function tells whether a year is a leap year or not.
}
\examples{
print(LeapYear(1990))
print(LeapYear(1991))
print(LeapYear(1992))
print(LeapYear(1993))
}
