% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dv-package.R
\docType{package}
\name{s2dv-package}
\alias{s2dv}
\alias{s2dv-package}
\title{s2dv: A Set of Common Tools for Seasonal to Decadal Verification}
\description{
The advanced version of package 's2dverification'. It is intended for 'seasonal to decadal' (s2d) climate forecast verification, but it can also be used in other kinds of forecasts or general climate analysis. This package is specially designed for the comparison between the experimental and observational datasets. The functionality of the included functions covers from data retrieval, data post-processing, skill scores against observation, to visualization. Compared to 's2dverification', 's2dv' is more compatible with the package 'startR', able to use multiple cores for computation and handle multi-dimensional arrays with a higher flexibility. The CDO version used in development is 1.9.8.
}
\references{
\url{https://earth.bsc.es/gitlab/es/s2dv/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://earth.bsc.es/gitlab/es/s2dv/}
  \item Report bugs at \url{https://earth.bsc.es/gitlab/es/s2dv/-/issues}
}

}
\author{
\strong{Maintainer}: Ariadna Batalla \email{ariadna.batalla@bsc.es} [contributor]

Authors:
\itemize{
  \item BSC-CNS [copyright holder]
  \item An-Chi Ho \email{an.ho@bsc.es}
  \item Nuria Perez-Zanon \email{nuria.perez@bsc.es}
}

Other contributors:
\itemize{
  \item Roberto Bilbao \email{roberto.bilbao@bsc.es} [contributor]
  \item Josep Cos \email{josep.cos@bsc.es} [contributor]
  \item Carlos Delgado \email{carlos.delgado@bsc.es} [contributor]
  \item Llorenç Lledó \email{llorenc.lledo@bsc.es} [contributor]
  \item Andrea Manrique \email{andrea.manrique@bsc.es} [contributor]
  \item Deborah Verfaillie \email{deborah.verfaillie@bsc.es} [contributor]
  \item Eva Rifà \email{eva.rifarovira@bsc.es} [contributor]
  \item Victòria Agudetse \email{victoria.agudetse@bsc.es} [contributor]
  \item Nadia Milders \email{nadia.milders@bsc.es} [contributor]
}

}
\keyword{internal}
