% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dv-package.R
\docType{package}
\name{s2dv-package}
\alias{s2dv}
\alias{s2dv-package}
\title{A Set of Common Tools for Seasonal to Decadal Verification}
\description{
The advanced version of package 's2dverification'. It is
intended for 'seasonal to decadal' (s2d) climate forecast verification, but
it can also be used in other kinds of forecasts or general climate analysis.
This package is specially designed for the comparison between the experimental
and observational datasets. The functionality of the included functions covers
from data retrieval, data post-processing, skill scores against observation,
to visualization. Compared to 's2dverification', 's2dv' is more compatible
with the package 'startR', able to use multiple cores for computation and
handle multi-dimensional arrays with a higher flexibility.
}
\references{
\url{https://earth.bsc.es/gitlab/es/s2dv/}
}
\keyword{internal}

